fixes broken saving of new UUID and progentor on clone, #257
